+++ CHANGELOG.md
@@ -3,6 +3,31 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com) and this project adheres to [Semantic Versioning](https://semver.org).
 
+## 2.0.0 - 2025-01-21
+
+### Added
+
+- Ability to merge external catalogue
+- phpunit/phpunit, orchestra/testbench dev dependency
+- citronel-general-error-catalogue config
+- citronel-error-config: citronel_error_catalogue_external_catalogues[] config to load external catalogues
+
+### Changed
+
+- citronel-error-catalogue
+
+### Deprecated
+
+- Nothing
+
+### Removed
+
+- Nothing
+
+### Fixed
+
+- Nothing
+
 ## 1.x.x - 2024-10-11
 
 ### Added